# openapi.yaml
openapi: 3.1.0
info:
  title: Async Agent API
  version: 0.1.0
  description: |
    RESTful API for managing autonomous agent goals, scheduled tasks, and run executions.
    Built with Fastify, uses SQLite for persistence, and supports multiple LLM providers.
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: http://localhost:3000
    description: Health check endpoints
security:
  - bearerAuth: []
  - {}
...
paths:
...
  /api/v1/goals:
    get:
      tags:
        - goals
      summary: List all goals
      description: Retrieve all goals with optional status filtering
      operationId: listGoals
      parameters:
        - $ref: '#/components/parameters/GoalStatusQuery'
      responses:
        '200':
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GoalWithSchedules'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - goals
      summary: Create a new goal
      description: Create a new goal with optional schedule configuration
      operationId: createGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
            examples:
              basic:
                summary: Basic goal
                value:
                  objective: Monitor GitHub repository for new issues and summarize them daily
                  params:
                    stepBudget: 20
              withSchedule:
                summary: Goal with schedule
                value:
                  objective: Daily market analysis
                  params:
                    stepBudget: 30
                    allowedTools:
                      - web_search
                      - web_scrape
                  schedule:
                    cronExpr: 0 9 * * *
                    timezone: UTC
      responses:
        '201':
          description: Goal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/goals/{id}:
    get:
      tags:
        - goals
      summary: Get goal by ID
      description: Retrieve a specific goal with its schedules
      operationId: getGoal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Goal details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalWithSchedules'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - goals
      summary: Update goal
      description: Update goal properties (objective, params, webhookUrl, status)
      operationId: updateGoal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalRequest'
            example:
              objective: Updated objective text
              status: paused
      responses:
        '200':
          description: Goal updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - goals
      summary: Delete goal
      description: Delete a goal and all its associated schedules and runs
      operationId: deleteGoal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '204':
          description: Goal deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/goals/{id}/run:
    post:
      tags:
        - goals
      summary: Trigger goal run
      description: Manually trigger a goal execution
      operationId: triggerGoalRun
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
            example: {}
      responses:
        '200':
          description: Run triggered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerRunResponse'
              example:
                runId: run_abc123
                message: Run triggered successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/goals/{id}/pause:
    post:
      tags:
        - goals
      summary: Pause goal
      description: Pause a goal and deactivate its schedules
      operationId: pauseGoal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
            example: {}
      responses:
        '200':
          description: Goal paused successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Goal paused successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/goals/{id}/resume:
    post:
      tags:
        - goals
      summary: Resume goal
      description: Resume a paused goal and reactivate its schedules
      operationId: resumeGoal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
            example: {}
      responses:
        '200':
          description: Goal resumed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Goal resumed successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/agents:
    get:
      tags:
        - agents
      summary: List all agents
      description: Retrieve all agents with optional filtering by name or active status
      operationId: listAgents
      parameters:
        - name: name
          in: query
          required: false
          description: Filter agents by name
          schema:
            type: string
        - name: active
          in: query
          required: false
          description: Filter agents by active status
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - agents
      summary: Create a new agent
      description: Register a new agent version with name, version, and prompt template
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Agent with this name and version already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/agents/{id}:
    get:
      tags:
        - agents
      summary: Get agent by ID
      description: Retrieve a specific agent by its ID
      operationId: getAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - agents
      summary: Delete agent
      description: Delete a specific agent (must not be active)
      operationId: deleteAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '204':
          description: Agent deleted successfully
        '400':
          description: Cannot delete active agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/agents/{id}/activate:
    post:
      tags:
        - agents
      summary: Activate an agent
      description: Activate an agent and deactivate other versions with the same name
      operationId: activateAgent
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
            example: {}
      responses:
        '200':
          description: Agent activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/agents/resolve/{name}:
    get:
      tags:
        - agents
      summary: Resolve active agent by name
      description: Get the currently active agent for a given name
      operationId: resolveAgent
      parameters:
        - name: name
          in: path
          required: true
          description: Agent name
          schema:
            type: string
      responses:
        '200':
          description: Active agent found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          description: No active agent found with this name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
  parameters:
    GoalId:
      name: id
      in: path
      required: true
      description: Goal ID (format goal_xxxxx)
      schema:
        type: string
        pattern: ^goal_[a-z0-9]+$
      example: goal_abc123
...
    AgentId:
      name: id
      in: path
      required: true
      description: Agent ID (format agent_xxxxx)
      schema:
        type: string
        pattern: ^agent_[a-z0-9]+$
      example: agent_abc123
    GoalStatusQuery:
      name: status
      in: query
      required: false
      description: Filter goals by status
      schema:
        $ref: '#/components/schemas/GoalStatus'
...
  schemas:
    GoalStatus:
      type: string
      enum:
        - active
        - paused
        - archived
      description: Status of a goal
...
    GoalParams:
      type: object
      properties:
        stepBudget:
          type: integer
          minimum: 1
          description: Maximum number of steps allowed
          example: 20
        allowedTools:
          type: array
          items:
            type: string
          description: List of tools the agent can use
          example:
            - web_search
            - web_scrape
        constraints:
          type: object
          additionalProperties: true
          description: Additional constraints for the goal
          example:
            maxRetries: 3
      additionalProperties: false
    Goal:
      type: object
      required:
        - id
        - objective
        - params
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          pattern: ^goal_[a-z0-9]+$
          description: Unique goal identifier
          example: goal_abc123
        objective:
          type: string
          minLength: 10
          maxLength: 1000
          description: The goal objective description
          example: Monitor GitHub repository for new issues
        params:
          $ref: '#/components/schemas/GoalParams'
        webhookUrl:
          type: string
          format: uri
          nullable: true
          description: Optional webhook URL for notifications
          example: https://example.com/webhook
        status:
          $ref: '#/components/schemas/GoalStatus'
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp
          example: '2025-10-30T10:30:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp
          example: '2025-10-30T10:30:00.000Z'
    Schedule:
      type: object
      required:
        - id
        - goalId
        - cronExpr
        - timezone
        - active
        - createdAt
      properties:
        id:
          type: string
          pattern: ^sched_[a-z0-9]+$
          description: Unique schedule identifier
          example: sched_xyz789
        goalId:
          type: string
          pattern: ^goal_[a-z0-9]+$
          description: Associated goal ID
          example: goal_abc123
        cronExpr:
          type: string
          description: Cron expression for scheduling
          example: 0 9 * * *
        timezone:
          type: string
          description: Timezone for schedule
          default: UTC
          example: UTC
        active:
          type: boolean
          description: Whether schedule is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp
          example: '2025-10-30T10:30:00.000Z'
    GoalWithSchedules:
      allOf:
        - $ref: '#/components/schemas/Goal'
        - type: object
          required:
            - schedules
          properties:
            schedules:
              type: array
              items:
                $ref: '#/components/schemas/Schedule'
              description: Associated schedules
...
    Agent:
      type: object
      required:
        - id
        - name
        - version
        - promptTemplate
        - active
        - metadata
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          pattern: ^agent_[a-z0-9]+$
          description: Unique agent identifier
          example: agent_abc123
        name:
          type: string
          description: Agent name
          example: defaultAgent
        version:
          type: string
          description: Agent version
          example: 1.0.0
        promptTemplate:
          type: string
          description: Agent prompt template
          example: You are a helpful assistant that...
        active:
          type: boolean
          description: Whether this agent version is active
          example: true
        metadata:
          type: object
          additionalProperties: true
          description: Additional agent metadata
          example: {}
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp
          example: '2025-10-30T10:30:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp
          example: '2025-10-30T10:30:00.000Z'
    CreateAgentRequest:
      type: object
      required:
        - name
        - version
        - promptTemplate
      properties:
        name:
          type: string
          minLength: 1
          description: Agent name
          example: defaultAgent
        version:
          type: string
          minLength: 1
          description: Agent version
          example: 1.0.0
        promptTemplate:
          type: string
          minLength: 1
          description: Agent prompt template
          example: You are a helpful assistant that...
        metadata:
          type: object
          additionalProperties: true
          description: Additional agent metadata
          example: {}
      additionalProperties: false
    CreateGoalRequest:
      type: object
      required:
        - objective
      properties:
        objective:
          type: string
          minLength: 10
          maxLength: 1000
          description: The goal objective description
          example: Monitor GitHub repository for new issues
        params:
          $ref: '#/components/schemas/GoalParams'
        webhookUrl:
          type: string
          format: uri
          description: Optional webhook URL for notifications
          example: https://example.com/webhook
        schedule:
          type: object
          required:
            - cronExpr
          properties:
            cronExpr:
              type: string
              description: Cron expression
              example: 0 9 * * *
            timezone:
              type: string
              description: Timezone for schedule
              default: UTC
              example: UTC
          additionalProperties: false
      additionalProperties: false
    UpdateGoalRequest:
      type: object
      properties:
        objective:
          type: string
          minLength: 10
          maxLength: 1000
          description: Updated goal objective
          example: Updated objective text
        params:
          $ref: '#/components/schemas/GoalParams'
        webhookUrl:
          type: string
          format: uri
          nullable: true
          description: Updated webhook URL
          example: https://example.com/webhook
        status:
          $ref: '#/components/schemas/GoalStatus'
      additionalProperties: false
...
    TriggerRunResponse:
      type: object
      required:
        - runId
        - message
      properties:
        runId:
          type: string
          pattern: ^run_[a-z0-9]+$
          description: ID of the triggered run
          example: run_abc123
        message:
          type: string
          example: Run triggered successfully
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Resource not found
  responses:
    BadRequest:
      description: Bad Request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Invalid request parameters
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Resource not found
    TooManyRequests:
      description: Too many requests - rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Rate limit exceeded. Maximum 100 requests per minute.
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error